% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_texture_points}
\alias{add_texture_points}
\title{Add \code{ggplot2::geom_point()} points to texture triangle}
\usage{
add_texture_points(df, sand, silt, clay, ...)
}
\arguments{
\item{df}{Data frame containing columns for sand, silt, and clay.}

\item{sand, silt, clay}{Column names in \code{df} for sand, silt, and clay \% values.}

\item{...}{Other arguments to pass into \code{ggplot2::aes()}.}
}
\value{
\code{ggplot2} texture triangle with points for all samples.
}
\description{
Add \code{ggplot2::geom_point()} points to texture triangle
}
\examples{
library(ggplot2)
# Create a texture triangle with points colored by texture
make_texture_triangle(body_font = "sans") +
  add_texture_points(
    df = washi_data,
    sand = sand_percent,
    silt = silt_percent,
    clay = clay_percent,
    color = "#a60f2d",
    alpha = 0.6
  ) +
  # if you are setting a color to a constant, you must call
  # scale_*_identity().
  ggplot2::scale_color_identity() +
  ggplot2::scale_alpha_identity()

# Remember these are `ggplot2` functions and require `+` instead of
#  pipes (`|>` or `\%>\%`)
try({
  make_texture_triangle(body_font = "sans") +
    add_texture_points(
      df = washi_data,
      sand = sand_percent,
      silt = silt_percent,
      clay = clay_percent,
      color = texture
    ) +
    ggplot2::scale_color_viridis_d()
})
}
