% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_mode}
\alias{calculate_mode}
\title{Calculate the mode of a categorical variable}
\usage{
calculate_mode(x)
}
\arguments{
\item{x}{Character vector to calculate the mode from.}
}
\value{
A single value (character) that occurs most often in \code{x}.
}
\description{
Returns the most frequently occurring value in a character vector, ignoring
missing values.
}
\examples{
calculate_mode(washi_data$crop)

}
