% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{check_additional_columns}
\alias{check_additional_columns}
\title{Check for presence of additional measurement columns}
\usage{
check_additional_columns(data)
}
\arguments{
\item{data}{Data frame of input data.}
}
\value{
A list of error issues.
}
\description{
Ensures that the data contains at least one column beyond the required fields.
}
\keyword{internal}
