% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{check_data_types}
\alias{check_data_types}
\title{Check data types against requirements}
\usage{
check_data_types(x)
}
\arguments{
\item{x}{Named list with \code{data} and \code{data_dict}.}
}
\value{
A list of error issues.
}
\description{
Ensures that column data types match expected types defined in
\code{required_fields}.
}
\keyword{internal}
