% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{check_dict_mismatch}
\alias{check_dict_mismatch}
\title{Check for mismatches between data and data dictionary}
\usage{
check_dict_mismatch(data, data_dict)
}
\arguments{
\item{data}{Data frame of input data.}

\item{data_dict}{Data frame of data dictionary.}
}
\value{
A list of warning issues.
}
\description{
Identifies discrepancies between columns in the data and entries in the
data dictionary.
}
\keyword{internal}
