% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{check_large_measurement_groups}
\alias{check_large_measurement_groups}
\title{Check for large measurement groups}
\usage{
check_large_measurement_groups(data_dict)
}
\arguments{
\item{data_dict}{Data dictionary.}
}
\value{
A list of warning issue objects.
}
\description{
Warns when a measurement group contains more than 8 measurements.
Large groups may cause report figures and tables to appear squished.
}
\keyword{internal}
