% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{check_numeric_conversion}
\alias{check_numeric_conversion}
\title{Detect non-numeric values in measurement columns}
\usage{
check_numeric_conversion(data, data_dict)
}
\arguments{
\item{data}{Data frame of input data.}

\item{data_dict}{Data frame of data dictionary.}
}
\value{
A list of warning issues.
}
\description{
Attempts numeric conversion and identifies values that are coerced to \code{NA}.
}
\keyword{internal}
