% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{check_uniqueness}
\alias{check_uniqueness}
\title{Check uniqueness constraints}
\usage{
check_uniqueness(x)
}
\arguments{
\item{x}{Named list with \code{data} and \code{data_dict}.}
}
\value{
A list of error issues.
}
\description{
Validates that specified fields meet uniqueness requirements,
either globally or within grouping variables.
}
\keyword{internal}
