% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip-plots.R
\name{convert_ggiraph}
\alias{convert_ggiraph}
\title{Convert a \code{ggplot2} plot to an interactive \code{ggiraph}}
\usage{
convert_ggiraph(plot, ..., body_font = "Poppins", width = 6, height = 4)
}
\arguments{
\item{plot}{\code{ggplot2} plot to convert to interactive \code{ggiraph}. \code{plot} must
contain \verb{ggiraph::geom_<plot_type>_interactive()}.}

\item{...}{Other arguments passed to \code{\link[ggiraph:girafe_options]{ggiraph::girafe_options()}}.}

\item{body_font}{Font family to use throughout plot. Defaults to
\code{"Poppins"}.}

\item{width}{Width of SVG output in inches. Defaults to 6.}

\item{height}{Height of SVG output in inches. Defaults to 4.}
}
\value{
Facetted strip plots with classes of \code{girafe} and \code{htmlwidget}.
}
\description{
Convert a \code{ggplot2} plot to an interactive \code{ggiraph}
}
\examples{
# Read in wrangled example plot data
df_plot_path <- soils_example("df_plot.RDS")
df_plot <- readRDS(df_plot_path)

# Make strip plot with all measurements and set scales based on
# the category column and then apply theme.

# Subset df to just biological measurement group
df_plot_bio <- df_plot |>
  dplyr::filter(measurement_group == "biological")

# NOTE: the plot gets piped into the `set_scales()` function, which gets
# added to `theme_facet_strip()`.
plot <- make_strip_plot(
  df_plot_bio,
  x = dummy,
  y = value,
  id = sample_id,
  group = abbr_unit,
  tooltip = label,
  color = category,
  fill = category,
  size = category,
  alpha = category,
  shape = category
) |>
  set_scales() +
  theme_facet_strip(body_font = "sans")

# Convert static plot to interactive `ggiraph`
convert_ggiraph(plot)
}
