% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_soils.R
\name{create_soils}
\alias{create_soils}
\title{Create a project directory for soil health reports using a Quarto template}
\usage{
create_soils(path, overwrite = FALSE, open = TRUE)
}
\arguments{
\item{path}{Name of project directory to be created.}

\item{overwrite}{Boolean. Overwrite the existing project?}

\item{open}{Boolean. Open the newly created project?}
}
\value{
A new project directory containing template and resources.
}
\description{
This function was adapted from
https://github.com/ThinkR-open/golem/blob/365a5cc303b189973abab0dd375c64be79bcf74a/R/create_golem.R
}
\examples{
\dontrun{
create_soils_project(tempdir("soils"))
}
}
