% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary}
\alias{data_dictionary}
\title{Data dictionary}
\format{
\subsection{\code{data_dictionary} A data frame with 32 rows and 7 columns.}{

\describe{
\item{measurement_group}{Name to group measurements by}
\item{measurement_group_label}{Label of measurement group to be used as
heading}
\item{column_name}{Name of column in data set, used for joining}
\item{order}{Order of how measurements are presented within each
measurement_group}
\item{abbr}{Abbreviated measurement name for labels}
\item{unit}{Measurement unit}
\item{abbr_unit}{HTML formatted abbreviation with unit for plots and
tables}
...
}
}
}
\source{
\href{https://washingtonsoilhealthinitiative.com/state-of-the-soils/}{WaSHI State of the Soils}
}
\usage{
data_dictionary
}
\description{
An example data dictionary for the Washington Soil Health Initiative (WaSHI)
State of the Soils Assessment anonymized data.
}
\keyword{datasets}
