% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary}
\alias{dictionary}
\title{Data dictionary}
\format{
\subsection{\code{dictionary} A data frame with 32 rows and 7 columns.}{

columns:
\describe{
\item{measurement_group}{Name to group measurements by}
\item{order}{Order of how measurements are presented within each measurement_group}
\item{column_name}{Name of column in dataset, used for joining}
\item{measurement_full_name}{Measurement fully spelled out}
\item{abbr}{Abbreviated measurement name for labels}
\item{unit}{Measurement unit}
\item{abbr_unit}{HTML formatted abbreviation with unit for plots and tables}
...
}
}
}
\source{
\url{https://agr.wa.gov/departments/land-and-water/natural-resources/soil-health/state-of-the-soils}
}
\usage{
dictionary
}
\description{
An example data dictionary for the example dataset.
}
\keyword{datasets}
