% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_long}
\alias{example_data_long}
\title{Example data in long (tidy) format}
\format{
\subsection{\code{example_data_long} A data frame with 1,800 rows and 14 columns:}{

\describe{
\item{year}{Year of sample}
\item{sampleId, producerId, fieldId}{Anonymized IDs}
\item{farmName, producerName, fieldName}{Anonymized names}
\item{longitude, latitude}{Truncated coordinates}
\item{texture}{Measured soil texture}
\item{measurement}{Measurement name with units}
\item{value}{Measurement result}
...
}
}
}
\source{
\url{https://agr.wa.gov/departments/land-and-water/natural-resources/soil-health/state-of-the-soils}
}
\usage{
example_data_long
}
\description{
A subset WaSHI State of the Soils Assessment dataset that has been
anonymized. This dataset is tidied, so each measurement is in its
own row.
}
\keyword{datasets}
