% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data_wide}
\alias{example_data_wide}
\title{Example data in wide format}
\format{
\subsection{\code{example_data_wide} A data frame with 30 rows and 72}{

columns:
\describe{
\item{year}{Year of sample}
\item{sampleId, producerId, fieldId}{Anonymized IDs}
\item{farmName, producerName, fieldName}{Anonymized names}
\item{longitude, latitude}{Truncated coordinates}
\item{texture}{Measured soil texture}
\item{other columns}{Column name includes measurement and units; value is the measurement results}
...
}
}
}
\source{
\url{https://agr.wa.gov/departments/land-and-water/natural-resources/soil-health/state-of-the-soils}
}
\usage{
example_data_wide
}
\description{
A subset WaSHI State of the Soils Assessment dataset that has been
anonymized. This dataset presents each sample in its own row, with
columns for each measurement.
}
\keyword{datasets}
