% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Example data}
\format{
\subsection{\code{exampleData} A data frame with 100 rows and 51 columns:}{

\describe{
\item{year}{Year of sample}
\item{sampleId, producerId, fieldId}{Anonymized IDs}
\item{farmName, producerName, fieldName}{Anonymized names}
\item{longitude, latitude}{Truncated coordinates}
\item{texture}{Measured soil texture}
\item{other columns}{Column name includes measurement and units; value is the measurement results}
...
}
}
}
\source{
\href{https://washingtonsoilhealthinitiative.com/state-of-the-soils/}{WaSHI State of the Soils}
}
\usage{
exampleData
}
\description{
A subset of the Washington Soil Health Initiative (WaSHI) State of the Soils
Assessment anonymized data. This data set presents each sample in its own
row, with columns for each measurement.
}
\keyword{datasets}
