% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{format_ft_colors}
\alias{format_ft_colors}
\title{Conditionally flextable background cell colors.}
\usage{
format_ft_colors(
  ft,
  lighter_col = washi::washi_pal[["standard"]][["cream"]],
  darker_col = washi::washi_pal[["standard"]][["tan"]]
)
}
\arguments{
\item{ft}{Flextable object}

\item{lighter_col}{Lighter background color. Defaults to WaSHI cream.}

\item{darker_col}{Darker background color. Defaults to WaSHI tan.}
}
\description{
Color the background cells based on how the value compares
to the project average (as long as the project average is the last
row of the table).
}
