% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_table_headers}
\alias{get_table_headers}
\title{Get table headers for flextable}
\usage{
get_table_headers(dictionary, group)
}
\arguments{
\item{dictionary}{Data frame containing columns \code{measurement_group}, \code{abbr},
\code{unit}.}

\item{group}{Character \code{measurement_group} value.}
}
\description{
Internal helper that uses the data dictionary to construct
flextable column headers for a single measurement group.
}
