% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{make_interactive_map}
\alias{make_interactive_map}
\title{Make an interactive map of soil sample locations with \code{leaflet}}
\source{
JavaScript code adapted from
\href{https://github.com/bhaskarvk/leaflet.extras/tree/master}{leaflet.extras}.
}
\usage{
make_interactive_map(df, primary_color = "#a60f2d")
}
\arguments{
\item{df}{Data frame containing \code{longitude}, \code{latitude}, \code{label}, and \code{popup}
columns. See \code{prep_for_map()} for details.}

\item{primary_color}{A character string specifying the color used for point
features on the map (hex code or color name). Defaults to WaSHI red
(#a60f2d).}
}
\value{
A \pkg{leaflet} map widget.
}
\description{
Create an interactive web map with point locations and popups using
\pkg{leaflet}. This function is designed to work with data prepared using
\code{prep_for_map()}.
}
\examples{
gis_df <- washi_data |>
  dplyr::distinct(latitude, longitude, .keep_all = TRUE) |>
  dplyr::select(c(latitude, longitude, farm_name, crop)) |>
  head(3) |>
  prep_for_map(label_heading = farm_name, label_body = crop)

dplyr::glimpse(gis_df)

make_interactive_map(gis_df)
}
\seealso{
\code{make_static_map()}
}
