% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_plotly}
\alias{make_plotly}
\title{make plotly}
\usage{
make_plotly(
  data,
  primary_color = washi_pal$green,
  secondary_color = washi_pal$gray,
  other_color = washi_pal$tan,
  primary_accent_color = washi_pal$blue
)
}
\arguments{
\item{data}{dataframe}
}
\description{
make plotly
}
