% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{make_static_map}
\alias{make_static_map}
\title{Make a static map of soil sample locations with \code{ggplot2}}
\usage{
make_static_map(
  df,
  label = field_id,
  provider = "Esri.WorldImagery",
  body_font = "sans",
  primary_color = "#a60f2d"
)
}
\arguments{
\item{df}{Data frame containing at minimum \code{longitude} and \code{latitude} columns.}

\item{label}{Name of the column in \code{df} used to label soil sample points.}

\item{provider}{Character string specifying the basemap tile provider, passed
to \code{maptiles::get_tiles()}. See details of \code{maptiles::get_tiles()} for
available providers. Defaults to "Esri.WorldImagery".}

\item{body_font}{Character string specifying the font family used for map
labels. Defaults to "sans".}

\item{primary_color}{A character string specifying the color used for point
features on the map (hex code or color name). Defaults to WaSHI red
(#a60f2d).}
}
\value{
A \code{{ggplot2}} object representing the static map.
}
\description{
Create a static map using basemap tiles and sample point locations. This
function is intended as a non-interactive alternative to
\code{make_interactive_map()}.
}
\examples{
gis_df <- washi_data |>
  dplyr::distinct(latitude, longitude, .keep_all = TRUE) |>
  dplyr::select(c(latitude, longitude, farm_name)) |>
  head(3)

dplyr::glimpse(gis_df)

\dontrun{
static_map <- make_static_map(gis_df, label = farm_name)
static_map
}
}
