% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_texture_triangle}
\alias{make_texture_triangle}
\title{Make a texture triangle}
\usage{
make_texture_triangle(body_font = "Poppins")
}
\arguments{
\item{body_font}{Font family to use throughout plot. Defaults to
\code{"Poppins"}.}
}
\value{
Blank \code{ggplot2} texture triangle with USDA textural classes.
}
\description{
Make a texture triangle
}
\examples{
# Blank texture triangle with just USDA classes
make_texture_triangle(body_font = "sans")

}
