% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pull_unique}
\alias{pull_unique}
\title{Extract unique values from a single column of a data frame}
\usage{
pull_unique(df, target)
}
\arguments{
\item{df}{A data frame containing the column of interest.}

\item{target}{Column to extract unique values from (string or unquoted
symbol).}
}
\value{
A vector of unique values from the specified column.
}
\description{
Extract unique values from a single column of a data frame
}
\examples{
pull_unique(washi_data, "crop")

}
