% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_docx}
\alias{render_docx}
\title{render_docx}
\usage{
render_docx(producerId)
}
\arguments{
\item{producerId}{}
}
\description{
render_docx
}
\examples{
unique(results22$producerId) |>
as.character() |>
walk(render_docx)
}
