% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_html}
\alias{render_html}
\title{render html}
\usage{
render_html(producerId)
}
\arguments{
\item{producerId}{}
}
\description{
render html
}
\examples{
unique(results22$producerId) |>
as.character() |>
walk(render_html)
}
