% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_producer_report}
\alias{render_producer_report}
\title{Render producer reports}
\usage{
render_producer_report(producerId, year, output = "html", output_dir = NULL)
}
\arguments{
\item{producerId}{Character \code{producerId} to render report for.}

\item{year}{Year of samples to include in report.}

\item{output}{Target output format.

Currently supported options:
\itemize{
\item \code{"html"} for an interactive report.
\item \code{"docx"} for an editable word document.
\item \code{"all"} for rendering all supported output formats.
}}

\item{output_dir}{Path to the directory that holds the reports. If \code{NULL},
the reports will be saved in the project directory.}
}
\description{
Render \verb{_producerReport.qmd} to \code{.html} or \code{.docx} files. The word document
outputs should be opened and manually edited to ensure the report looks good.
See \code{vignette("docx", package = "soils")} for more details on these edits
}
\examples{
\dontrun{
# Render single html report for first producer in data set
first_producer <- head(exampleData$producerId, 1)
render_producer_report(
  first_producer,
  year = 2023,
  output = "html",
  output_dir = paste0(here::here(),"/inst/reports/")
)

# Render docx reports for all 2023 producers
unique(exampleData$producerId) |>
  purrr::walk(
    \(producerId) render_producer_report(
      producerId,
      year = 2023,
      output = "docx",
      output_dir = paste0(here::here(),"/inst/reports/"),
      .progress = TRUE
    )
  )
}
}
