% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_report}
\alias{render_report}
\title{Render reports}
\usage{
render_report(producerId, year, input, output = "html", output_dir = NULL)
}
\arguments{
\item{producerId}{Character \code{producerId} to render report for.}

\item{year}{Year of samples to include in report.}

\item{input}{Input \code{.qmd} file name. Must be within the \code{inst} folder.}

\item{output}{Target output format.

Currently supported options:
\itemize{
\item \code{"html"} for an interactive report.
\item \code{"docx"} for an editable word document.
}}

\item{output_dir}{Path to the directory that holds the reports. If \code{NULL},
the reports will be saved in the project directory.}
}
\description{
Wrapper for \code{quarto::quarto_render()} with additional project specific
arguments. Limits output formats to \code{.html} or \code{.docx}. The word document
outputs should be opened and manually edited to ensure the report looks good.
See \code{vignette("docx", package = "soils")} for more details on these edits
}
\examples{
\dontrun{
# Render single html report for first producer in data set
first_producer <- head(exampleData$producerId, 1)
render_report(
  first_producer,
  year = 2023,
  input = "producer_report.qmd",
  output = "html",
  output_dir = paste0(here::here(), "/inst/reports/")
)

producers <- exampleData |>
  subset(year == 2023) |>
  dplyr::distinct(producerId)

# Render docx for these 2023 producers to the `/inst/reports/` directory
purrr::walk(
  producers$producerId,
  \(producerId) render_report(
    producerId,
    year = 2023,
    input = "producer_report.qmd",
    output = "docx",
    output_dir = paste0(here::here(), "/inst/reports/")
  ),
  .progress = TRUE
)
}
}
