% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_to_html}
\alias{render_to_html}
\title{Render HTML reports}
\usage{
render_to_html(producerId, year)
}
\arguments{
\item{producerId}{Character producerId to render report for.}

\item{year}{Year of samples to include in report.}
}
\description{
Unfortunately, self-contained html reports must be rendered to the same folder
as the .qmd html reports should be manually moved to the inst/reports/
subfolder after rendering
https://stackoverflow.com/questions/72346829/two-problems-rendering-a-qmd-file-with-quarto-render-from-r
}
\examples{
\dontrun{
unique(example_data_wide$producerId) |>
  as.character() |>
  purrr::walk(\(x) render_to_html(x, 2022),
    .progress = TRUE
  )
}
}
