% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{set_scales}
\alias{set_scales}
\title{Define styles for producer's samples versus all samples}
\usage{
set_scales(
  plot,
  primary_color = washi::washi_pal[["standard"]][["red"]],
  secondary_color = washi::washi_pal[["standard"]][["ltgray"]],
  other_color = washi::washi_pal[["standard"]][["tan"]]
)
}
\arguments{
\item{plot}{\code{ggplot} object to apply scales to.}

\item{primary_color}{Color of producer's sample points. Defaults to WaSHI
green.}

\item{secondary_color}{Color of sample points with \code{"Same crop"} or \code{"Same county"} values in the \code{category} column. Defaults to WaSHI gray.}

\item{other_color}{Color of sample points with \code{"Other fields"} value in
\code{category} column. Defaults to WaSHI tan.}
}
\value{
\code{ggplot} object with manual alpha, color, shape, and size scales
applied.
}
\description{
Define styles for producer's samples versus all samples
}
\examples{
# Read in wrangled example plot data
df_plot_path <- soils_example("df_plot.RDS")
df_plot <- readRDS(df_plot_path)
# Subset df to just biological measurement group
df_plot_bio <- df_plot |>
  subset(measurement_group == "biological")

# Make strip plot

make_strip_plot(
  df_plot_bio,
  x = dummy,
  y = value,
  id = sampleId,
  group = abbr_unit,
  tooltip = label,
  color = category,
  size = category,
  alpha = category,
  shape = category
) |>
  set_scales() +
  theme_facet_strip(body_font = "sans")

# Example without setting scales
make_strip_plot(
  df_plot_bio,
  x = dummy,
  y = value,
  id = sampleId,
  group = abbr_unit,
  tooltip = label
) +
  theme_facet_strip(body_font = "sans")
}
