% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils-package.R
\docType{package}
\name{soils-package}
\alias{soils}
\alias{soils-package}
\title{soils: Visualize and Report Soil Health Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Collection of soil health data visualization and reporting tools, including a RStudio project template with everything you need to generate custom HTML and Microsoft Word reports for each participant in your soil health survey.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WA-Department-of-Agriculture/soils}
  \item \url{https://wa-department-of-agriculture.github.io/soils/}
  \item Report bugs at \url{https://github.com/WA-Department-of-Agriculture/soils/issues}
}

}
\author{
\strong{Maintainer}: Jadey Ryan \email{jryan@agr.wa.gov}

Authors:
\itemize{
  \item Molly McIlquham
  \item Kwabena A Sarpong
  \item Leslie Michel
  \item Teal Potter
  \item Deirdre Griffin LaHue
  \item Dani L Gelardi
}

Other contributors:
\itemize{
  \item Washington State Department of Agriculture [copyright holder, funder]
}

}
\keyword{internal}
