% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{soils_example}
\alias{soils_example}
\title{Get path to example data}
\source{
Adapted from \code{readxl::readxl_example()}.
}
\usage{
soils_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
\code{soils} comes bundled with some example files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
soils_example()

soils_example("df_plot.RDS")
}
