% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{split_issues}
\alias{split_issues}
\title{Split validation issues into errors and warnings}
\usage{
split_issues(issues)
}
\arguments{
\item{issues}{List of issues created by new_issue().}
}
\value{
Named list with $errors and $warnings.
}
\description{
Split validation issues into errors and warnings
}
