% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{style_ft}
\alias{style_ft}
\title{style flextable}
\usage{
style_ft(ft, header_color = washi_pal$green, border_color = washi_pal$tan)
}
\arguments{
\item{ft}{flextable object}
}
\description{
style flextable
}
