% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summarize_by_project}
\alias{summarize_by_project}
\title{Summarize samples across the project}
\usage{
summarize_by_project(results_long)
}
\arguments{
\item{results_long}{Data frame in tidy, long format with columns: \code{sample_id},
\code{measurement_group}, \code{abbr}, \code{value}.}
}
\value{
A data frame summarizing the average values across the project by
\code{measurement_group} and \code{abbr}. Includes a "Field or Average" column and,
if present, a \code{Texture} column.
}
\description{
Summarize samples across the project
}
