% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usda_texture}
\alias{usda_texture}
\title{USDA texture data}
\format{
\subsection{\code{usda_texture} A list of two data frames: \code{polygons}}{

contains 53 rows and 4 columns and \code{labels} contains 12 rows and 5 columns.
\describe{
\item{clay}{Percentage of clay content}
\item{sand}{Percentage of sand content}
\item{silt}{Percentage of silt content}
\item{label}{Texture class}
\item{angle}{Angle to rotate the label in the texture triangle}
...
}
}
}
\source{
Adapted from \code{data(USDA, package = "ggtern")}.
}
\usage{
usda_texture
}
\description{
US Department of Agriculture textural classes
}
\keyword{datasets}
