% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{washi_data}
\alias{washi_data}
\title{Example WaSHI data}
\format{
\subsection{\code{washi_data} A data frame with 100 rows and 42 columns:}{

\describe{
\item{year}{Year of sample}
\item{sample_id, producer_id, field_id}{Anonymized IDs}
\item{farm_name, field_name}{Anonymized names}
\item{longitude, latitude}{Truncated coordinates}
\item{texture}{Measured soil texture}
\item{other columns}{Column name includes measurement and units;
value is the measurement result}
...
}
}
}
\source{
\href{https://washingtonsoilhealthinitiative.com/state-of-the-soils/}{WaSHI State of the Soils}
}
\usage{
washi_data
}
\description{
A subset of the Washington Soil Health Initiative (WaSHI) State of the Soils
Assessment anonymized data. This data set presents each sample in its own
row, with columns for each measurement.
}
\keyword{datasets}
