% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mode.R
\name{calculate_mode}
\alias{calculate_mode}
\title{calculate mode of categorical variable}
\usage{
calculate_mode(x)
}
\arguments{
\item{x}{character vector}
}
\description{
calculate mode of categorical variable
}
