% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_mode}
\alias{calculate_mode}
\title{Calculate the mode of categorical variable}
\usage{
calculate_mode(x)
}
\arguments{
\item{x}{Character vector to calculate mode from.}
}
\value{
The value that occurred most often.
}
\description{
Calculate the mode of categorical variable
}
\examples{
calculate_mode(washi_data$crop)

}
