% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{check_measurement_groups}
\alias{check_measurement_groups}
\title{Validate measurement group values}
\usage{
check_measurement_groups(data_dict, lanugage = c("English", "Spanish"))
}
\arguments{
\item{data_dict}{Data frame of data dictionary.}

\item{language}{Character. \code{"English"} or \code{"Spanish"}.}
}
\value{
A list of warning issues.
}
\description{
Ensures that \code{measurement_group} values match expected controlled
vocabularies.
}
\keyword{internal}
