% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_soils.R
\name{create_soils}
\alias{create_soils}
\title{Create a project directory for soil health reports using a Quarto template}
\source{
Adapted from \code{golem::create_golem()}.
}
\usage{
create_soils(path, overwrite = FALSE, open = TRUE)
}
\arguments{
\item{path}{Name of project directory to be created.}

\item{overwrite}{Boolean. Overwrite the existing project?}

\item{open}{Boolean. Open the newly created project?}
}
\value{
A new project directory containing template and resources.
}
\description{
Creates an RStudio project containing Quarto template and resources (images,
style sheets, etc.) and R scripts with functions used in the Quarto template
for your modification. See \code{vignettes("project", "soils")} for more details.
}
\examples{
\dontrun{
# Create temporary directory
dir <- tempdir()

# Create soils project
create_soils(dir, overwrite = TRUE)

# Delete temporary directory
unlink(dir, recursive = TRUE)
}
}
