% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-soils.R
\name{create_soils}
\alias{create_soils}
\title{Create a project directory for generating soil health reports}
\source{
Adapted from \code{golem::create_golem()}.
}
\usage{
create_soils(path, template = "English", overwrite = FALSE, open = TRUE)
}
\arguments{
\item{path}{Name of project directory to be created.}

\item{template}{Template type. Either "English" or "Spanish".}

\item{overwrite}{Boolean. Overwrite the existing project?}

\item{open}{Boolean. Open the newly created project?}
}
\value{
A new project directory containing template and resources.
}
\description{
Creates an RStudio project containing Quarto template and resources (images,
style sheets, render.R script).
}
\examples{
\dontrun{
# Create temporary directory
dir <- tempdir()

# Create soils project
create_soils(dir, overwrite = TRUE)

# Delete temporary directory
unlink(dir, recursive = TRUE)
}
}
