% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Example data}
\format{
\subsection{\code{exampleData} A data frame with 100 rows and 51}{

columns:
\describe{
\item{year}{Year of sample}
\item{sampleId, producerId, fieldId}{Anonymized IDs}
\item{farmName, producerName, fieldName}{Anonymized names}
\item{longitude, latitude}{Truncated coordinates}
\item{texture}{Measured soil texture}
\item{other columns}{Column name includes measurement and units; value is the measurement results}
...
}
}
}
\source{
\url{https://agr.wa.gov/departments/land-and-water/natural-resources/soil-health/state-of-the-soils}
}
\usage{
exampleData
}
\description{
A subset WaSHI State of the Soils Assessment data set that has been
anonymized. This data set presents each sample in its own row, with
columns for each measurement.
}
\keyword{datasets}
