% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{format_ft_colors}
\alias{format_ft_colors}
\title{Conditional formatting of flextable background cell colors}
\usage{
format_ft_colors(
  ft,
  lighter_color = "#F2F0E6",
  darker_color = "#ccc29c",
  language = "English"
)
}
\arguments{
\item{ft}{Flextable object}

\item{lighter_color}{Lighter background color. Defaults to WaSHI cream.}

\item{darker_color}{Darker background color. Defaults to WaSHI tan.}

\item{language}{Language of the footnote. "English" (default) or "Spanish".}
}
\description{
Color the background cells based on how the value compares to the project
average. The project average must be the last row of the table. A footnote is
added to the table describing what the dark and light colors mean.
}
\examples{
# Read in wrangled example table data
tables_path <- soils_example("tables.RDS")
tables <- readRDS(tables_path)

# Make the table
ft <- flextable::flextable(tables$biological)
ft

# Conditionally format background cell colors
format_ft_colors(ft)
}
