% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_n_texture_by_var}
\alias{get_n_texture_by_var}
\title{Calculate n samples and most frequent texture by a grouping variable}
\usage{
get_n_texture_by_var(results_long, producer_info, var)
}
\arguments{
\item{results_long}{Dataframe in tidy, long format with columns: \code{sample_id},
\code{texture}.}

\item{producer_info}{Vector of producer's values for the grouping variable.}

\item{var}{Variable to group and summarize by.}
}
\description{
This function is used in \code{summarize_by_var}.
}
