% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_table_headers}
\alias{get_table_headers}
\title{Get table headers for flextable}
\usage{
get_table_headers(dictionary, group)
}
\arguments{
\item{dictionary}{Dataframe containing columns \code{measurement_group}, \code{abbr},
\code{unit}.}

\item{group}{Character \code{measurement_group} value.}
}
\description{
This function uses the data dictionary to create a new dataframe of the
abbreviations and units for each measurement group for flextable
}
