% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_column_empty}
\alias{is_column_empty}
\title{Check if a column is empty (all NA or blank)}
\usage{
is_column_empty(column)
}
\arguments{
\item{column}{to check if all values are NA or empty strings}
}
\description{
Check if a column is empty (all NA or blank)
}
