% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_ft}
\alias{make_ft}
\title{Make flextable for specified 'measurement_group'.}
\usage{
make_ft(measurement_group)
}
\arguments{
\item{measurement_group}{Name of measurement group to visualize in
flextable.}
}
\description{
This function requires the functions in 'prepare_data.R' to be run
first.
}
