% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_ft}
\alias{make_ft}
\title{Make a flextable with column names from another dataframe}
\usage{
make_ft(table, header)
}
\arguments{
\item{table}{A dataframe with the contents of the desired flextable output.}

\item{header}{Another dataframe with three columns:
\itemize{
\item First column contains what the top header row should be. In our template,
this is the abbreviation of the measurement (i.e. \verb{Organic Matter}).
\item Second column, called \code{"key"}, contains the join key. In our template,
this is the same as the first column.
\item Third column contains the second header row. In our template, this is
the unit (i.e. \verb{\%}).
}}
}
\value{
Formatted flextable object.
}
\description{
Make a flextable with column names from another dataframe
}
\examples{
# Read in wrangled table data
headers_path <- soils_example("headers.RDS")
headers <- readRDS(headers_path)

tables_path <- soils_example("tables.RDS")
tables <- readRDS(tables_path)

# Input dataframes
headers$chemical

tables$chemical

# Make the flextable
make_ft(
  table = tables$chemical,
  header = headers$chemical
) |>
  # Style the flextable
  style_ft() |>
  # Add the white line under the columns with the same units
  unit_hline(header = headers$chemical)

}
