% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{make_leaflet}
\alias{make_leaflet}
\title{Make leaflet map}
\usage{
make_leaflet(df, primary_color = washi::washi_pal[["standard"]][["red"]])
}
\arguments{
\item{df}{Dataframe that contains sample info (sampleid, fieldid, crop, coordinates).}

\item{primary_color}{Color of points. Defaults to WaSHI red.}
}
\value{
Leaflet map.
}
\description{
Reset view code from https://github.com/bhaskarvk/leaflet.extras/blob/master/R/mapUtils.R
}
