% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{make_leaflet}
\alias{make_leaflet}
\title{Make leaflet map}
\source{
JavaScript code adapted from
\href{https://github.com/bhaskarvk/leaflet.extras/tree/master}{leaflet.extras}.
}
\usage{
make_leaflet(df, primary_color = "#a60f2d")
}
\arguments{
\item{df}{Dataframe containing columns: \code{longitude}, \code{latitude}, \code{label},
\code{popup}. See \code{prep_for_map()} for details.}

\item{primary_color}{Color of points. Defaults to WaSHI red.}
}
\value{
Leaflet map.
}
\description{
Make leaflet map
}
\examples{
gis_df <- washi_data |>
  dplyr::distinct(latitude, longitude, .keep_all = TRUE) |>
  head(3) |>
  prep_for_map(label_heading = field_name, label_body = crop)

dplyr::glimpse(gis_df)

# Make leaflet
make_leaflet(gis_df)
}
