% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_strip_plot}
\alias{make_strip_plot}
\title{make facetted strip plot}
\usage{
make_strip_plot(
  data,
  output,
  primary_color = washi_pal$green,
  secondary_color = washi_pal$gray,
  other_color = washi_pal$tan,
  primary_accent_color = washi_pal$blue
)
}
\arguments{
\item{data}{dataframe}

\item{output}{type of ouput, either "static" or "html"}
}
\description{
make facetted strip plot
}
