% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_strip_plot}
\alias{make_strip_plot}
\title{Make facetted strip plot}
\usage{
make_strip_plot(
  df,
  output,
  font_family = "Poppins",
  primary_color = washi::washi_pal[["standard"]][["green"]],
  secondary_color = washi::washi_pal[["standard"]][["gray"]],
  other_color = washi::washi_pal[["standard"]][["tan"]],
  primary_accent_color = washi::washi_pal[["standard"]][["blue"]]
)
}
\arguments{
\item{df}{Dataframe containing columns: 'category', 'abbr_unit',
'unit', 'dummy', 'sampleLabel'.}

\item{output}{Type of ouput: either "static" or "html"}

\item{font_family}{Font family to use throughout plot. Defaults to
"Poppins".}

\item{primary_color}{Color of producer's sample points. Defaults to
WaSHI green.}

\item{secondary_color}{Color of sample points with 'sameCrop' or
sameCounty' categories. Defaults to WaSHI gray.}

\item{other_color}{Color of all other sample points in project.
Defaults to WaSHI tan.}

\item{primary_accent_color}{Color of facet strip background.
Defaults to WaSHI blue.}
}
\description{
Make facetted strip plot
}
