% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip-plots.R
\name{make_strip_plot}
\alias{make_strip_plot}
\title{Make a facetted strip plot}
\usage{
make_strip_plot(
  df,
  ...,
  x = dummy,
  y = value,
  id = sample_id,
  group = abbr_unit,
  tooltip = label,
  language = "English"
)
}
\arguments{
\item{df}{Data frame to plot.}

\item{...}{Other arguments passed to \code{\link[graphics:points]{graphics::points()}}.}

\item{x}{Column for x-axis. For these strip plots, we recommend using a dummy
variable to act as a placeholder. Defaults to a column named \code{dummy} with
only one value ("dummy") for all rows.}

\item{y}{Column for y-axis. Defaults to \code{value}.}

\item{id}{Column with unique identifiers for each sample to use as \code{data_id}
for interactive plots. Defaults to \code{sample_id}.}

\item{group}{Column to facet by. Defaults to \code{abbr_unit}.}

\item{tooltip}{Column with tooltip labels for interactive plots.}

\item{language}{Language of the legend. \code{"English"} (default) or \code{"Spanish"}.}
}
\value{
Facetted \code{ggplot2} strip plots.
}
\description{
Make a facetted strip plot
}
\examples{
# Read in wrangled example plot data
df_plot_path <- soils_example("df_plot.RDS")
df_plot <- readRDS(df_plot_path)

# Subset df to just biological measurement group
df_plot_bio <- df_plot |>
  dplyr::filter(measurement_group == "biological")

# Make strip plot with all measurements and set scales based on
# the category column and then apply theme.

# NOTE: the plot gets piped into the `set_scales()` function, which gets
# added to `theme_facet_strip()`.

make_strip_plot(
  df_plot_bio,
  x = dummy,
  y = value,
  id = sample_id,
  group = abbr_unit,
  tooltip = label,
  color = category,
  fill = category,
  size = category,
  alpha = category,
  shape = category
) |>
  set_scales() +
  theme_facet_strip(body_font = "sans")

# Example of strip plot without scales or theme functions
make_strip_plot(df_plot_bio)

# Example of strip plot with `x` set to the facet group instead of a dummy
# variable. The dummy variable is what centers the points within the subplot.
make_strip_plot(
  df_plot_bio,
  x = abbr_unit,
  y = value,
  id = sample_id,
  group = abbr_unit,
  tooltip = label,
  color = category,
  fill = category,
  size = category,
  alpha = category,
  shape = category
) |>
  set_scales() +
  theme_facet_strip(body_font = "sans")
}
