% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texture-triangle.R
\name{make_texture_triangle}
\alias{make_texture_triangle}
\title{Make a textural class triangle}
\source{
Adapted from {plotrix}:
\url{https://github.com/plotrix/plotrix/blob/0d4c2b065e2c2d327358ac8cdc0b0d46b89bea7f/R/soil.texture.R}
}
\usage{
make_texture_triangle(
  body_font = "Poppins",
  show_names = TRUE,
  show_lines = TRUE,
  show_grid = FALSE
)
}
\arguments{
\item{body_font}{Font family to use throughout plot. Defaults to \code{"Poppins"}.}

\item{show_names}{Boolean. Defaults to \code{TRUE} to show USDA textural class
names.}

\item{show_lines}{Boolean. Defaults to \code{TRUE} to show boundaries of USDA
textural classes.}

\item{show_grid}{Boolean. Defaults to \code{FALSE} to hide grid lines at each 10
level of each soil component.}
}
\value{
Opens the graphics device with a triangle plot containing USDA
textural classes.
}
\description{
Make a texture triangle with USDA textural classes.
}
\examples{
# Note the text appears squished in this example since the width, height,
# and resolution have been optimized to print the figure 6 in wide in the
# report.

make_texture_triangle(body_font = "sans")
}
