% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_samples.R
\name{map_samples}
\alias{map_samples}
\title{make leaflet map}
\usage{
map_samples(data, primary_color)
}
\arguments{
\item{data}{dataframe that contains sample info (sampleid, fieldid, crop, coordinates)}

\item{primary_color}{color of points}
}
\description{
reset view code from https://github.com/bhaskarvk/leaflet.extras/blob/master/R/mapUtils.R
}
