% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{new_issue}
\alias{new_issue}
\title{Create a validation issue}
\usage{
new_issue(severity = c("error", "warning"), message)
}
\arguments{
\item{severity}{"error" or "warning"}

\item{message}{A character string describing the issue}
}
\value{
A named list with severity and message.
}
\description{
Create a validation issue
}
