% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{prep_for_map}
\alias{prep_for_map}
\title{Prep data to gis df}
\usage{
prep_for_map(df, label_heading, label_body)
}
\arguments{
\item{df}{Dataframe containing columns: \code{longitude}, \code{latitude}, and two
columns with values you want to appear in the map label and popup.}

\item{label_heading}{Column in \code{df} that you want to appear as the bold point
label on your map, as well as the first line of the popup when the user
clicks a point.}

\item{label_body}{Column in \code{df} that you want to appear as body text below
the \code{label_heading} in the popup.}
}
\value{
Dataframe to be input into \code{make_leaflet()}.
}
\description{
Prep data to gis df
}
\examples{
washi_data |>
  dplyr::distinct(latitude, longitude, .keep_all = TRUE) |>
  head(3) |>
  prep_for_map(label_heading = field_name, label_body = crop) |>
  dplyr::glimpse()
}
