% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_unique.R
\name{pull_unique}
\alias{pull_unique}
\title{Pull unique values from one column of dataframe}
\usage{
pull_unique(data, target)
}
\arguments{
\item{data}{dataframe with column to extract unique values from}

\item{target}{variable to pull unique vector of (crop or county)}
}
\description{
Pull unique values from one column of dataframe
}
