% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pull_unique}
\alias{pull_unique}
\title{Pull unique values from one column of dataframe}
\usage{
pull_unique(df, target)
}
\arguments{
\item{df}{Dataframe with column to extract unique values from.}

\item{target}{Variable to pull unique vector of (i.e. crop or
county).}
}
\value{
Vector of unique values from target column.
}
\description{
Pull unique values from one column of dataframe
}
\examples{
washi_data |>
  pull_unique(crop)

}
