% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-soils-input.R
\name{read_soils_csv}
\alias{read_soils_csv}
\title{Read CSV input files}
\usage{
read_soils_csv(file)
}
\arguments{
\item{file}{Character vector of length 2.}
}
\value{
A named list with:
\describe{
\item{data}{A data frame containing the data file, or \code{NULL} if unreadable.}
\item{data_dict}{A data frame containing the data dictionary file, or \code{NULL} if unreadable.}
\item{issues}{A list of validation issue objects. Empty if no issues were found.}
}
}
\description{
Reads two CSV files: one containing the data and one containing the
data dictionary. Performs file name and readability validation checks.
}
\keyword{internal}
