% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_report}
\alias{render_report}
\title{Render producer reports}
\usage{
render_report(producerId, year, output = "html")
}
\arguments{
\item{producerId}{Character producerId to render report for.}

\item{year}{Year of samples to include in report.}

\item{output}{Target output format.

Currently supported options:
\itemize{
\item 'html' for an interactive report.
\item 'docx' for an editable word document.
\item 'all' for rendering all supported output formats.
}}
}
\description{
Render \verb{_producerReport.qmd} to .html or .docx files. The word document
outputs should be opened and manually tweaked to ensure the report looks
good. See \code{vignette("docx", package = "soils")} for more details on these
tweaks.
}
\examples{
\dontrun{
# Render single html report for first producer in data set
first_producer <- head(exampleData$producerId, 1)
render_report(
  first_producer,
  year = 2023,
  output = "html"
)

# Render docx reports for all 2023 producers
unique(exampleData$producerId) |>
  purrr::walk(
    \(producerId) render_report(
      producerId,
      year = 2023,
      output = "docx",
      .progress = TRUE
    )
  )
}
}
