% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_to_docx}
\alias{render_to_docx}
\title{Render word docs}
\usage{
render_to_docx(producerId, year)
}
\arguments{
\item{producerId}{Character producerId to render report for.}

\item{year}{Year of samples to include in report.}
}
\description{
After rendering to word, open files and make sure formatting looks good.
Flextables may need to be adjusted if there are many columns.
}
\examples{
\dontrun{
unique(example_data_wide$producerId) |>
  as.character() |>
  purrr::walk(\(x) render_to_html(x, 2022),
    .progress = TRUE
  )
}
}
