% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_to_html}
\alias{render_to_html}
\title{Render .qmd to html}
\usage{
render_to_html(producerId, year)
}
\arguments{
\item{producerId}{Character producerId to render report for.}

\item{year}{Year of samples to include in report.}
}
\description{
Render .qmd to html
}
\examples{
\dontrun{
unique(example_data_wide$producerId) |>
  as.character() |>
  purrr::walk(\(x) render_to_html(x, 2022))
}
}
