% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{save_plot}
\alias{save_plot}
\title{save plot}
\usage{
save_plot(plot, ext, height, width)
}
\arguments{
\item{plot}{name of plot to save}

\item{ext}{file extension type ("png", "svg", etc)}

\item{height}{height of plot}
}
\description{
save plot
}
