% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{save_plot}
\alias{save_plot}
\title{Save plot}
\usage{
save_plot(plot, ext, height, width)
}
\arguments{
\item{plot}{Name of plot to save.}

\item{ext}{File extension type ("png", "svg", etc).}

\item{height}{Height of plot.}

\item{width}{Width of plot.}
}
\value{
Side effects of saving plot.
}
\description{
Save plot
}
