% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip-plots.R
\name{set_scales}
\alias{set_scales}
\title{Define styles for producer's samples versus all samples}
\usage{
set_scales(
  plot,
  primary_color = "#a60f2d",
  secondary_color = "#3E3D3D",
  other_color = "#ccc29c",
  language = "English"
)
}
\arguments{
\item{plot}{\code{ggplot} object to apply scales to.}

\item{primary_color}{Color of producer's sample points. Defaults to WaSHI
red}

\item{secondary_color}{Color of sample points with \code{"Same crop"} or \code{"Same county"} values in the \code{category} column. Defaults to WaSHI gray.}

\item{other_color}{Color of sample points with \code{"Other fields"} value in
\code{category} column. Defaults to WaSHI tan.}

\item{language}{Language of the legend. \code{"English"} (default) or \code{"Spanish"}.}
}
\value{
\code{ggplot} object with manual alpha, color, shape, and size scales
applied.
}
\description{
Define styles for producer's samples versus all samples
}
\examples{
# Read in wrangled example plot data
df_plot_path <- soils_example("df_plot.RDS")
df_plot <- readRDS(df_plot_path)
# Subset df to just biological measurement group
df_plot_bio <- df_plot |>
  dplyr::filter(measurement_group == "biological")

# Make strip plot
make_strip_plot(
  df_plot_bio,
  x = dummy,
  y = value,
  id = sample_id,
  group = abbr_unit,
  tooltip = label,
  color = category,
  fill = category,
  size = category,
  alpha = category,
  shape = category
) |>
  set_scales() +
  theme_facet_strip(body_font = "sans")

# Example without setting scales
make_strip_plot(
  df_plot_bio,
  x = dummy,
  y = value,
  id = sample_id,
  group = abbr_unit,
  tooltip = label
) +
  theme_facet_strip(body_font = "sans")
}
