% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils-package.R
\docType{package}
\name{soils-package}
\alias{soils}
\alias{soils-package}
\title{soils: Visualize and Report Soil Health Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Collection of soil health data visualization and reporting tools, including a RStudio project template with everything you need to generate custom HTML and Microsoft Word reports for each participant in your soil health sampling project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/WA-Department-of-Agriculture/soils}
  \item \url{https://wa-department-of-agriculture.github.io/soils/}
  \item Report bugs at \url{https://github.com/WA-Department-of-Agriculture/soils/issues}
}

}
\author{
\strong{Maintainer}: Jadey N Ryan \email{jryan@agr.wa.gov} (\href{https://orcid.org/0009-0004-5998-4697}{ORCID})

Authors:
\itemize{
  \item Molly McIlquham (\href{https://orcid.org/0000-0002-0152-969X}{ORCID})
  \item Kwabena A Sarpong
  \item Leslie M Michel
  \item Teal S Potter (\href{https://orcid.org/0000-0002-4884-2720}{ORCID})
  \item Deirdre Griffin LaHue (\href{https://orcid.org/0000-0001-5711-797X}{ORCID})
  \item Dani L Gelardi (\href{https://orcid.org/0000-0002-5015-8119}{ORCID})
}

Other contributors:
\itemize{
  \item Washington State Department of Agriculture [copyright holder, funder]
}

}
\keyword{internal}
