% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-validation.R
\name{split_issues}
\alias{split_issues}
\title{Split validation issues by severity}
\usage{
split_issues(issues)
}
\arguments{
\item{issues}{List of issues created by \code{new_issue()}.}
}
\value{
A named list with:
\describe{
\item{errors}{List of error issues}
\item{warnings}{List of warning issues}
}
}
\description{
Separates a list of issues into errors and warnings.
}
\keyword{internal}
