% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{style_ft}
\alias{style_ft}
\title{Style a flextable}
\usage{
style_ft(
  ft,
  header_font = "Lato",
  body_font = "Poppins",
  header_color = washi::washi_pal[["standard"]][["green"]],
  header_text_color = "white",
  border_color = washi::washi_pal[["standard"]][["gray"]]
)
}
\arguments{
\item{ft}{Flextable object.}

\item{header_font}{Font of header text. Defaults to \code{"Lato"}.}

\item{body_font}{Font of body text. Defaults to \code{"Poppins"}.}

\item{header_color}{Background color of header cells. Defaults to WaSHI
green.}

\item{header_text_color}{Color of header text. Defaults to white.}

\item{border_color}{Color of border lines. Defaults to WaSHI gray.}
}
\value{
Styled flextable object.
}
\description{
Style a flextable
}
\examples{
# Read in wrangled example table data
# See `data_wrangling.R` for processing steps
tables_path <- soils_example("tables.RDS")
tables <- readRDS(tables_path)

# Make the table
ft <- flextable::flextable(tables$biological)
ft

# Style the table
style_ft(ft)
}
