% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summarize_by_project}
\alias{summarize_by_project}
\title{Summarize samples across the project}
\usage{
summarize_by_project(results_long)
}
\arguments{
\item{results_long}{Dataframe in tidy, long format with columns: \code{sample_id},
\code{texture}, \code{measurement_group}, \code{abbr}, \code{value}.}
}
\description{
Summarize samples across the project
}
