% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{summarize_by_var}
\alias{summarize_by_var}
\title{Summarize producer's samples with averages by grouping variable}
\usage{
summarize_by_var(results_long, producer_samples, var)
}
\arguments{
\item{results_long}{Dataframe in tidy, long format with columns: \code{sample_id},
\code{measurement_group}, \code{abbr}, \code{value}.}

\item{producer_samples}{Dataframe in tidy, long format with columns:
\code{measurement_group}, \code{abbr}, \code{value}.}

\item{var}{Variable to summarize by.}
}
\value{
A data frame summarizing the average values by measurement group,
\code{abbr}, and the specified grouping variable. Includes a "Field or Average"
column and, if present, a \code{Texture} column.
}
\description{
Summarize producer's samples with averages by grouping variable
}
