% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texture_triangle.R
\name{texture_triangle}
\alias{texture_triangle}
\title{Texture triangle}
\usage{
texture_triangle(
  data,
  primary_color = washi_pal$green,
  secondary_color = washi_pal$gray,
  other_color = washi_pal$tan
)
}
\arguments{
\item{data}{dataframe with columns: category, Clay, Sand, Silt}
}
\description{
Texture triangle
}
