% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{unit_hline}
\alias{unit_hline}
\title{Add bottom border to specific columns in flextable}
\usage{
unit_hline(ft, columns)
}
\arguments{
\item{ft}{flextable object}

\item{columns}{Indices of columns that were merged.}
}
\value{
Flextable object with bottom borders added.
}
\description{
Use when columns with the same units are merged together.
}
