% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{unit_hline}
\alias{unit_hline}
\title{unit_hline}
\usage{
unit_hline(ft, columns)
}
\arguments{
\item{ft}{flextable object}

\item{columns}{indices of columns that were merged to add border under header}
}
\value{
flextable object
}
\description{
unit_hline
}
