% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{unit_hline}
\alias{unit_hline}
\title{Add bottom border to specific columns in flextable}
\usage{
unit_hline(ft, columns)
}
\arguments{
\item{ft}{flextable object}

\item{columns}{Indices of columns that were merged.}
}
\value{
Flextable object with bottom borders added.
}
\description{
Use when columns with the same units are merged together to add a bottom
border to make it more obvious those columns share units.
}
\examples{
# Read in wrangled table data.
# See `data_wrangling.R` for processing steps.
headers_path <- soils_example("headers.RDS")
headers <- readRDS(headers_path)

tables_path <- soils_example("tables.RDS")
tables <- readRDS(tables_path)

# Make the table
ft <- make_ft(table = tables$chemical, header = headers$chemical)
ft

# Add a line under the merged columns with the same units
unit_hline(ft, 5)
}
